+++ cpp/open3d/visualization/gui/Menu.cpp
@@ -69,6 +69,14 @@ void Menu::InsertSeparator(int index) { impl_->menu->InsertSeparator(index); }
 
 int Menu::GetNumberOfItems() const { return impl_->menu->GetNumberOfItems(); }
 
+std::shared_ptr<MenuBase> Menu::GetMenu(const char* name) {
+    return impl_->menu->GetMenu(name);
+}
+
+std::shared_ptr<MenuBase> Menu::GetMenu(ItemId item_id) {
+    return impl_->menu->GetMenu(item_id);
+}
+
 bool Menu::IsEnabled(ItemId item_id) const {
     return impl_->menu->IsEnabled(item_id);
 }